Fixed directory_manager, fixed croatia metadata file

+++ sample_data/directory_manager.sh
@@ -6,12 +6,14 @@ for file in *.*json; do
   country="${file%%.*json}"
   country="${country%%_*}"
   echo "Working on ${country}..."
-  mkdir ./${country}/
+  mkdir -p ./${country}/
   mv ${file} ./${country}/${file}
   echo -e "# Sources\n" >> ./${country}/${country}_metadata.md
   echo -e "## ${file}\n\n\n" >> ./${country}/${country}_metadata.md
   for csv in ${country}*.csv; do
-    mv ${csv} ./${country}/${csv}
-    echo -e "### ${csv}\n\n\n" >> ./${country}/${country}_metadata.md
+    if [ -f "${csv}" ]; then
+      mv ${csv} ./${country}/${csv}
+      echo -e "### ${csv}\n\n\n" >> ./${country}/${country}_metadata.md
+    fi
   done
 done
